streamlit/requirements.txt
streamlit
google-generativeai>=0.8.3
numpy
langchain-google-genai
langchain-neo4j